+++ katarl2/agents/ppo/models/__init__.py
@@ -3,6 +3,7 @@ from .cnn_discrete import (
     Agent_IN, Agent_IN_before_norm,
     Agent_LN, Agent_LN_before_norm,
 )
+from .cnn_discrete_resnet import Agent as Agent_CNN_ResNet
 from .mlp_continuous import Agent as Agent_MLP_continuous
 from .mlp_continuous_simba import Agent as Agent_MLP_continuous_Simba
 from .cnn_discrete_simba import Agent as Agent_CNN_Simba