+++ katarl2/agents/ppo/ppo_env_cfg.py
@@ -1,8 +1,9 @@
 import numpy as np
 from functools import partial
 from typing import Optional, Any
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from katarl2.envs import GymAtariEnvConfig, GymMujocoEnvConfig, DMCEnvConfig, EnvpoolAtariEnvConfig
+from katarl2.envs.common.env_cfg import AtariWrapperConfig
 
 @dataclass
 class PPOEnvpoolAtariEnvConfig(EnvpoolAtariEnvConfig):
@@ -26,6 +27,46 @@ class PPOGymMujocoEnvConfig(GymMujocoEnvConfig):
 @dataclass
 class PPODMCEnvConfig(DMCEnvConfig):
     num_envs: int = 1
+
+    normalize_observation: bool = True
+    transform_observation: Optional[Any] = partial(np.clip, a_min=-10, a_max=10)
+    clip_action: bool = True
+    normalize_reward: bool = True
+    normalize_reward_gamma: float = 0.99
+    transform_reward: Optional[Any] = partial(np.clip, a_min=-10, a_max=10)
+
+
+@dataclass
+class SPOEnvpoolAtariEnvConfig(EnvpoolAtariEnvConfig):
+    num_envs: int = 8
+    atari_wrapper_cfg: AtariWrapperConfig = field(
+        default_factory=lambda: AtariWrapperConfig(episodic_life=True)
+    )
+
+
+@dataclass
+class SPOGymAtariEnvConfig(GymAtariEnvConfig):
+    num_envs: int = 8
+    atari_wrapper_cfg: AtariWrapperConfig = field(
+        default_factory=lambda: AtariWrapperConfig(episodic_life=True)
+    )
+
+
+@dataclass
+class SPOGymMujocoEnvConfig(GymMujocoEnvConfig):
+    num_envs: int = 8
+
+    normalize_observation: bool = True
+    transform_observation: Optional[Any] = partial(np.clip, a_min=-10, a_max=10)
+    clip_action: bool = True
+    normalize_reward: bool = True
+    normalize_reward_gamma: float = 0.99
+    transform_reward: Optional[Any] = partial(np.clip, a_min=-10, a_max=10)
+
+
+@dataclass
+class SPODMCEnvConfig(DMCEnvConfig):
+    num_envs: int = 8
 
     normalize_observation: bool = True
     transform_observation: Optional[Any] = partial(np.clip, a_min=-10, a_max=10)